+++ lumin/nn/training/fold_train.py
@@ -37,7 +37,7 @@ def get_folds(val_idx, n_folds, shuffle_folds:bool=True):
 def fold_train_ensemble(fy:FoldYielder, n_models:int, bs:int, model_builder:ModelBuilder,
                         callback_args:Dict[str,Dict[str,Any]]={}, eval_metrics:Dict[str,EvalMetric]={},
                         train_on_weights:bool=True, eval_on_weights:bool=True, patience:int=10, max_epochs:int=200,
-                        plots:List[str]=['history'], shuffle_fold:bool=True, shuffle_folds:bool=True, bulk_move:bool=True,
+                        plots:List[str]=['history', 'realtime'], shuffle_fold:bool=True, shuffle_folds:bool=True, bulk_move:bool=True,
                         savepath:Path=Path('train_weights'), verbose:bool=False, log_output:bool=False,
                         plot_settings:PlotSettings=PlotSettings()) -> Tuple[List[Dict[str,float]],List[Dict[str,List[float]]],List[Dict[str,float]]]:
     '''Train a specified numer of models created by ModelBuilder on data provided by FoldYielder, and save them to savepath'''
@@ -54,7 +54,7 @@ def fold_train_ensemble(fy:FoldYielder, n_models:int, bs:int, model_builder:Mode
     nb = len(fy.foldfile['fold_0/targets'])//bs
 
     model_bar = master_bar(np.random.choice(range(fy.n_folds), size=n_models, replace=False))
-    model_bar.names = ['Best', 'Train', 'Validation']
+    if 'realtime' in plots: model_bar.names = ['Best', 'Train', 'Validation']
     for model_num, val_id in enumerate(model_bar):
         print(f"Training model {model_num+1} / {n_models}")
         model_tmr = timeit.default_timer()
@@ -86,9 +86,9 @@ def fold_train_ensemble(fy:FoldYielder, n_models:int, bs:int, model_builder:Mode
             c.on_train_begin()
 
         val_x, val_y, val_w = Tensor(val_fold['inputs']), Tensor(val_fold['targets']), to_tensor(val_fold['weights']) if train_on_weights else None
-        model_bar.update_graph([[0, 0] for i in range(len(model_bar.names))])
+        if 'realtime' in plots: model_bar.update_graph([[0, 0] for i in range(len(model_bar.names))])
         epoch_pb = progress_bar(range(max_epochs), leave=True)
-        model_bar.show()
+        if 'realtime' in plots: model_bar.show()
         for epoch in epoch_pb:
             for trn_id in trn_ids:
                 subEpoch += 1
@@ -125,7 +125,7 @@ def fold_train_ensemble(fy:FoldYielder, n_models:int, bs:int, model_builder:Mode
                     epoch_counter += 1
 
                 x = np.arange(len(loss_history['val_loss']))
-                model_bar.update_graph([[x, best_loss*np.ones_like(x)]] + [[x, loss_history[l]] for l in loss_history])
+                if 'realtime' in plots: model_bar.update_graph([[x, best_loss*np.ones_like(x)]] + [[x, loss_history[l]] for l in loss_history])
 
                 if epoch_counter >= patience or model.stop_train:  # Early stopping
                     print('Early stopping after {} epochs'.format(subEpoch))
@@ -147,7 +147,7 @@ def fold_train_ensemble(fy:FoldYielder, n_models:int, bs:int, model_builder:Mode
         with open(savepath/'results_file.pkl', 'wb') as fout: pickle.dump(results, fout)
         with open(savepath/'cycle_file.pkl', 'wb') as fout: pickle.dump(cycle_losses, fout)
         
-        delattr(model_bar, 'fig')
+        if 'realtime' in plots: delattr(model_bar, 'fig')
         plt.clf()
         if 'cycle' in plots and cyclic_callback is not None: cyclic_callback.plot()
         print(f"Fold took {timeit.default_timer()-model_tmr:.3f}s\n")